Add package-wide test utils in tests/__init__.py
